+++ src/client/game.h
@@ -397,7 +397,7 @@ private:
     bool m_denyBotCall{ false };
     bool m_dead{ false };
     bool m_expertPvpMode;
-    int m_serverBeat{ 50 };
+    uint16_t m_serverBeat{ 50 };
     ticks_t m_ping{ -1 };
     uint32_t m_pingSent{ 0 };
     uint32_t m_pingReceived{ 0 };
@@ -405,7 +405,7 @@ private:
 
     stdext::timer m_pingTimer;
     Timer m_dashTimer;
-    int m_pingDelay{ 1000 };
+    uint16_t m_pingDelay{ 1000 };
     Otc::FightModes m_fightMode{ Otc::FightBalanced };
     Otc::ChaseModes m_chaseMode{ Otc::DontChase };
     Otc::PVPModes m_pvpMode{ Otc::WhiteDove };
@@ -423,10 +423,10 @@ private:
     ScheduledEventPtr m_walkEvent;
     ScheduledEventPtr m_checkConnectionEvent;
     bool m_connectionFailWarned;
-    int m_protocolVersion{ 0 };
-    int m_clientVersion{ 0 };
-    int m_lastSupportedVersion{ 1287 };
-    int m_openPvpSituations{ 0 };
+    uint16_t m_protocolVersion{ 0 };
+    uint16_t m_clientVersion{ 0 };
+    uint16_t m_lastSupportedVersion{ 1287 };
+    uint8_t m_openPvpSituations{ 0 };
     std::string m_clientSignature;
     Otc::OperatingSystem_t m_clientCustomOs{ Otc::CLIENTOS_NONE };
 };
